requirements.txt
pandas
numpy
sentence-transformers
chromadb
google-genai
python-dotenv
streamlit
plotly
torch